requirements.txt
streamlit
pandas
scikit-learn
pdfplumber
python-docx
nltk
sentence-transformers
torch
